Amazing-scripts/requirements.txt
pytesseract==0.2.6
Pillow==6.2.0
fitz==0.0.1.dev2
